EN-3395 :fix after review

+++ statusHandler/termuiStatusHandler.go
@@ -22,6 +22,7 @@ func NewTermuiStatusHandler() *TermuiStatusHandler {
 	return tsh
 }
 
+// StartTermuiConsole method will start termui console
 func (tsh *TermuiStatusHandler) StartTermuiConsole() error {
 	err := tsh.tui.Start()
 
@@ -37,32 +38,32 @@ func (tsh *TermuiStatusHandler) Termui() *termuic.TermuiConsole {
 func (tsh *TermuiStatusHandler) initMetricsMap() {
 	tsh.termuiConsoleMetrics = &sync.Map{}
 
-	tsh.termuiConsoleMetrics.Store(core.MetricNonce, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricCurrentRound, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricIsSyncing, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricNumConnectedPeers, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricSynchronizedRound, 0)
+	tsh.termuiConsoleMetrics.Store(core.MetricNonce, uint64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricCurrentRound, int64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricIsSyncing, uint64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricNumConnectedPeers, int64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricSynchronizedRound, uint64(0))
 
 	tsh.termuiConsoleMetrics.Store(core.MetricPublicKey, "")
-	tsh.termuiConsoleMetrics.Store(core.MetricShardId, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricTxPoolLoad, 0)
+	tsh.termuiConsoleMetrics.Store(core.MetricShardId, uint64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricTxPoolLoad, int64(0))
 
-	tsh.termuiConsoleMetrics.Store(core.MetricCountConsensus, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricCountLeader, 0)
-	tsh.termuiConsoleMetrics.Store(core.MetricCountAcceptedBlocks, 0)
+	tsh.termuiConsoleMetrics.Store(core.MetricCountConsensus, uint64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricCountLeader, uint64(0))
+	tsh.termuiConsoleMetrics.Store(core.MetricCountAcceptedBlocks, uint64(0))
 }
 
 // SetInt64Value method - will update the value for a key
 func (tsh *TermuiStatusHandler) SetInt64Value(key string, value int64) {
 	if _, ok := tsh.termuiConsoleMetrics.Load(key); ok {
-		tsh.termuiConsoleMetrics.Store(key, int(value))
+		tsh.termuiConsoleMetrics.Store(key, value)
 	}
 }
 
 // SetUInt64Value method - will update the value for a key
 func (tsh *TermuiStatusHandler) SetUInt64Value(key string, value uint64) {
 	if _, ok := tsh.termuiConsoleMetrics.Load(key); ok {
-		tsh.termuiConsoleMetrics.Store(key, int(value))
+		tsh.termuiConsoleMetrics.Store(key, value)
 	}
 }
 
@@ -77,7 +78,7 @@ func (tsh *TermuiStatusHandler) SetStringValue(key string, value string) {
 func (tsh *TermuiStatusHandler) Increment(key string) {
 	if keyValueI, ok := tsh.termuiConsoleMetrics.Load(key); ok {
 
-		keyValue := keyValueI.(int)
+		keyValue := keyValueI.(uint64)
 		keyValue++
 		tsh.termuiConsoleMetrics.Store(key, keyValue)
 	}
